# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Restaurante CRUD",
        "description": "API para la gestión de Restaurantes: Comensales, Mesas y Reservas",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/Comensal/crudComensal": {
            "post": {
                "tags": [
                    "Comensal"
                ],
                "summary": "Crear, editar o eliminar un comensal",
                "description": "Dependiendo del valor del parámetro 'action', se puede crear (c), actualizar (u) o eliminar (d) un comensal.",
                "operationId": "a20c93d406b1da33c0df49c7515e7383",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "action"
                                ],
                                "properties": {
                                    "action": {
                                        "description": "Acción a realizar: 'c' para crear, 'u' para actualizar, 'd' para eliminar",
                                        "type": "string",
                                        "example": "c"
                                    },
                                    "id": {
                                        "description": "ID del comensal (requerido para 'u' y 'd')",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "description": "Nombre del comensal (requerido para 'c' y 'u')",
                                        "type": "string",
                                        "example": "JOSE ZAFRA"
                                    },
                                    "correo": {
                                        "description": "Correo electrónico del comensal",
                                        "type": "string",
                                        "example": "josereynaldozv@gmail.com"
                                    },
                                    "telefono": {
                                        "description": "Teléfono del comensal",
                                        "type": "string",
                                        "example": "934466163"
                                    },
                                    "direccion": {
                                        "description": "Dirección del comensal",
                                        "type": "string",
                                        "example": "AV. BALTA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operación exitosa"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "JOSE ZAFRA"
                                                },
                                                "correo": {
                                                    "type": "string",
                                                    "example": "josereynaldozv@gmail.com"
                                                },
                                                "telefono": {
                                                    "type": "string",
                                                    "example": "934466163"
                                                },
                                                "direccion": {
                                                    "type": "string",
                                                    "example": "AV. BALTA"
                                                },
                                                "state": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-04-21T14:22:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-04-21T14:22:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        },
                                        "data": {
                                            "properties": {
                                                "error_message": {
                                                    "type": "string",
                                                    "example": "Detalle del error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}